@@ -1,16 +1,8 @@
 import { Injectable } from '@angular/core';
-import { Actions, Effect, ofType } from '@ngrx/effects';
-import {
-  DeleteRotation,
-  GetRotation,
-  MyRotationsLoaded,
-  RotationLoaded,
-  RotationPersisted,
-  RotationsActionTypes,
-  UpdateRotation
-} from './rotations.actions';
+import { Actions, createEffect, ofType } from '@ngrx/effects';
+import { DeleteRotation, GetRotation, MyRotationsLoaded, RotationLoaded, RotationPersisted, RotationsActionTypes, UpdateRotation } from './rotations.actions';
 import { AuthFacade } from '../../../+state/auth.facade';
-import { catchError, exhaustMap, filter, map, mergeMap, switchMap, withLatestFrom, distinctUntilChanged } from 'rxjs/operators';
+import { catchError, distinctUntilChanged, exhaustMap, filter, map, mergeMap, switchMap, withLatestFrom } from 'rxjs/operators';
 import { CraftingRotationService } from '../../../core/database/crafting-rotation/crafting-rotation.service';
 import { TeamcraftUser } from '../../../model/user/teamcraft-user';
 import { EMPTY, of } from 'rxjs';
@@ -18,8 +10,8 @@ import { EMPTY, of } from 'rxjs';
 @Injectable()
 export class RotationsEffects {
 
-  @Effect()
-  loadMyRotations$ = this.actions$.pipe(
+
+  loadMyRotations$ = createEffect(() => this.actions$.pipe(
     ofType(RotationsActionTypes.LoadMyRotations),
     switchMap(() => this.authFacade.userId$),
     distinctUntilChanged(),
@@ -28,10 +20,10 @@ export class RotationsEffects {
         map(rotations => new MyRotationsLoaded(rotations, userId))
       );
     })
-  );
+  ));
 
-  @Effect()
-  getRotation$ = this.actions$.pipe(
+
+  getRotation$ = createEffect(() => this.actions$.pipe(
     ofType<GetRotation>(RotationsActionTypes.GetRotation),
     mergeMap(action => {
       return this.rotationsService.get(action.key).pipe(
@@ -39,10 +31,10 @@ export class RotationsEffects {
       );
     }),
     map(rotation => new RotationLoaded(rotation))
-  );
+  ));
 
-  @Effect()
-  updateRotation$ = this.actions$.pipe(
+
+  updateRotation$ = createEffect(() => this.actions$.pipe(
     ofType<UpdateRotation>(RotationsActionTypes.UpdateRotation),
     withLatestFrom(this.authFacade.userId$),
     mergeMap(([action, userId]) => {
@@ -57,16 +49,16 @@ export class RotationsEffects {
     }),
     filter(res => res !== null),
     switchMap((res: string) => of(new RotationPersisted(res)))
-  );
+  ));
 
-  @Effect()
-  deleteRotation$ = this.actions$.pipe(
+
+  deleteRotation$ = createEffect(() => this.actions$.pipe(
     ofType<DeleteRotation>(RotationsActionTypes.DeleteRotation),
     mergeMap((action) => {
       return this.rotationsService.remove(action.key);
     }),
     switchMap(() => EMPTY)
-  );
+  ), { dispatch: false });
 
   constructor(
     private actions$: Actions,
