@@ -1,12 +1,12 @@
 import { DataModel } from '../database/storage/data-model';
 import { ForeignKey } from '../database/relational/foreign-key';
 import { TeamcraftUser } from '../../model/user/teamcraft-user';
-import { Aetheryte } from '../data/aetheryte';
 import { Vector3 } from '../tools/vector3';
 import { FishingBait } from '../data/model/fishing-bait';
 import { Hookset } from '../data/model/hookset';
 import { SpearfishingSpeed } from '../data/model/spearfishing-speed';
 import { SpearfishingShadowSize } from '../data/model/spearfishing-shadow-size';
+import { LazyAetheryte } from '../../lazy-data/model/lazy-aetheryte';
 
 export class Alarm extends DataModel {
 
@@ -67,7 +67,7 @@ export class Alarm extends DataModel {
 
   note: string;
 
-  aetheryte?: Aetheryte;
+  aetheryte?: LazyAetheryte;
 
   folklore?: number;
 
