@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { Actions, Effect, ofType } from '@ngrx/effects';
+import { Actions, createEffect, ofType } from '@ngrx/effects';
 import {
   CreateCustomItem,
   CreateCustomItemFolder,
@@ -20,8 +20,8 @@ import { CustomItemFoldersService } from '../custom-item-folders.service';
 @Injectable()
 export class CustomItemsEffects {
 
-  @Effect()
-  loadCustomItems$ = this.actions$.pipe(
+  
+  loadCustomItems$ = createEffect(() => this.actions$.pipe(
     ofType(CustomItemsActionTypes.LoadCustomItems),
     switchMap(() => this.authFacade.userId$),
     distinctUntilChanged(),
@@ -31,10 +31,10 @@ export class CustomItemsEffects {
           map(items => new CustomItemsLoaded(items))
         );
     })
-  );
+  ));
 
-  @Effect({ dispatch: false })
-  createCustomItem$ = this.actions$.pipe(
+  
+  createCustomItem$ = createEffect(() => this.actions$.pipe(
     ofType<CreateCustomItem>(CustomItemsActionTypes.CreateCustomItem),
     withLatestFrom(this.authFacade.userId$),
     switchMap(([action, userId]) => {
@@ -45,28 +45,28 @@ export class CustomItemsEffects {
         return this.customItemsService.set(action.payload.$key, action.payload).pipe(map(() => action.payload.$key));
       }
     })
-  );
+  ), { dispatch: false });
 
 
-  @Effect({ dispatch: false })
-  updateCustomItem$ = this.actions$.pipe(
+  
+  updateCustomItem$ = createEffect(() => this.actions$.pipe(
     ofType<UpdateCustomItem>(CustomItemsActionTypes.UpdateCustomItem),
     switchMap(action => {
       delete action.payload.folderId;
       delete action.payload.dirty;
       return this.customItemsService.update(action.payload.$key, action.payload);
     })
-  );
+  ), { dispatch: false });
 
 
-  @Effect({ dispatch: false })
-  deleteCustomItem$ = this.actions$.pipe(
+  
+  deleteCustomItem$ = createEffect(() => this.actions$.pipe(
     ofType<DeleteCustomItem>(CustomItemsActionTypes.DeleteCustomItem),
     switchMap(action => this.customItemsService.remove(action.key))
-  );
+  ), { dispatch: false });
 
-  @Effect()
-  loadCustomItemFolders$ = this.actions$.pipe(
+  
+  loadCustomItemFolders$ = createEffect(() => this.actions$.pipe(
     ofType(CustomItemsActionTypes.LoadCustomItemFolders),
     switchMap(() => this.authFacade.userId$),
     distinctUntilChanged(),
@@ -76,31 +76,31 @@ export class CustomItemsEffects {
           map(folders => new CustomItemFoldersLoaded(folders))
         );
     })
-  );
+  ));
 
-  @Effect({ dispatch: false })
-  createCustomItemFolder$ = this.actions$.pipe(
+  
+  createCustomItemFolder$ = createEffect(() => this.actions$.pipe(
     ofType<CreateCustomItemFolder>(CustomItemsActionTypes.CreateCustomItemFolder),
     withLatestFrom(this.authFacade.userId$),
     switchMap(([action, userId]) => {
       action.payload.authorId = userId;
       return this.customItemFoldersService.add(action.payload);
     })
-  );
+  ), { dispatch: false });
 
 
-  @Effect({ dispatch: false })
-  updateCustomItemFolder$ = this.actions$.pipe(
+  
+  updateCustomItemFolder$ = createEffect(() => this.actions$.pipe(
     ofType<UpdateCustomItemFolder>(CustomItemsActionTypes.UpdateCustomItemFolder),
     switchMap(action => this.customItemFoldersService.update(action.payload.$key, action.payload))
-  );
+  ), { dispatch: false });
 
 
-  @Effect({ dispatch: false })
-  deleteCustomItemFolder$ = this.actions$.pipe(
+  
+  deleteCustomItemFolder$ = createEffect(() => this.actions$.pipe(
     ofType<DeleteCustomItemFolder>(CustomItemsActionTypes.DeleteCustomItemFolder),
     switchMap(action => this.customItemFoldersService.remove(action.key))
-  );
+  ), { dispatch: false });
 
   constructor(
     private actions$: Actions,
