@@ -2,7 +2,7 @@ import { CustomItem } from '../../custom-items/model/custom-item';
 import { ListRow } from './list-row';
 
 export interface CraftAddition {
-  item: ListRow;
+  item?: ListRow;
   data?: CustomItem;
   amount: number;
 }
