@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { Actions, Effect, ofType } from '@ngrx/effects';
+import { Actions, createEffect, ofType } from '@ngrx/effects';
 import {
   CreateRotationFolder,
   DeleteRotationFolder,
@@ -23,8 +23,8 @@ import { RotationsFacade } from '../../rotations/+state/rotations.facade';
 @Injectable()
 export class RotationFoldersEffects {
 
-  @Effect()
-  loadMyRotationFolders$ = this.actions$.pipe(
+
+  loadMyRotationFolders$ = createEffect(() => this.actions$.pipe(
     ofType(RotationFoldersActionTypes.LoadMyRotationFolders),
     switchMap(() => this.authFacade.userId$),
     switchMap(userId => {
@@ -32,10 +32,10 @@ export class RotationFoldersEffects {
         map(folders => new MyRotationFoldersLoaded(folders, userId))
       );
     })
-  );
+  ));
 
-  @Effect()
-  loadRotationFolder$ = this.actions$.pipe(
+
+  loadRotationFolder$ = createEffect(() => this.actions$.pipe(
     ofType<LoadRotationFolder>(RotationFoldersActionTypes.LoadRotationFolder),
     mergeMap(action => {
       return this.craftingRotationFolderService.get(action.key).pipe(
@@ -48,10 +48,10 @@ export class RotationFoldersEffects {
       return folder.rotationIds.forEach(rotationId => this.rotationsFacade.getRotation(rotationId));
     }),
     map(folder => new RotationFolderLoaded(folder))
-  );
+  ));
 
-  @Effect()
-  createRotationFolder$ = this.actions$.pipe(
+
+  createRotationFolder$ = createEffect(() => this.actions$.pipe(
     ofType<CreateRotationFolder>(RotationFoldersActionTypes.CreateRotationFolder),
     withLatestFrom(this.authFacade.userId$),
     switchMap(([action, userId]) => {
@@ -62,26 +62,24 @@ export class RotationFoldersEffects {
       return this.craftingRotationFolderService.add(action.folder);
     }),
     map((key: string) => new FolderCreated(key))
-  );
+  ));
 
 
-  @Effect()
-  updateRotationFolder$ = this.actions$.pipe(
+  updateRotationFolder$ = createEffect(() => this.actions$.pipe(
     ofType<UpdateRotationFolder>(RotationFoldersActionTypes.UpdateRotationFolder),
     mergeMap(action => this.craftingRotationFolderService.set(action.folder.$key, action.folder)),
     switchMap(() => EMPTY)
-  );
+  ), { dispatch: false });
 
 
-  @Effect()
-  deleteRotationFolder$ = this.actions$.pipe(
+  deleteRotationFolder$ = createEffect(() => this.actions$.pipe(
     ofType<DeleteRotationFolder>(RotationFoldersActionTypes.DeleteRotationFolder),
     mergeMap(action => this.craftingRotationFolderService.remove(action.key)),
     switchMap(() => EMPTY)
-  );
+  ), { dispatch: false });
 
-  @Effect()
-  removeListFromWorkshop$ = this.actions$.pipe(
+
+  removeListFromWorkshop$ = createEffect(() => this.actions$.pipe(
     ofType<RemoveRotationFromFolder>(RotationFoldersActionTypes.RemoveRotationFromFolder),
     withLatestFrom(this.foldersFacade.allRotationFolders$),
     map(([action, folders]: [RemoveRotationFromFolder, CraftingRotationsFolder[]]) => {
@@ -89,7 +87,7 @@ export class RotationFoldersEffects {
       folder.rotationIds = folder.rotationIds.filter(id => id !== action.rotationKey);
       return new UpdateRotationFolder(folder);
     })
-  );
+  ));
 
   constructor(
     private actions$: Actions,
