@@ -1,28 +1,28 @@
 import { Injectable } from '@angular/core';
-import { Actions, Effect, ofType } from '@ngrx/effects';
+import { Actions, createEffect, ofType } from '@ngrx/effects';
 import { CreateTeam, DeleteTeam, LoadMyTeams, LoadTeam, MyTeamsLoaded, TeamLoaded, TeamsActionTypes, UpdateTeam } from './teams.actions';
 import { AuthFacade } from '../../../+state/auth.facade';
 import { catchError, distinctUntilChanged, exhaustMap, filter, map, switchMap, withLatestFrom } from 'rxjs/operators';
 import { TeamService } from '../../../core/database/team.service';
-import { EMPTY, of } from 'rxjs';
+import { of } from 'rxjs';
 import { Team } from '../../../model/team/team';
 import { TeamsFacade } from './teams.facade';
 
 @Injectable()
 export class TeamsEffects {
 
-  @Effect()
-  loadMyTeams$ = this.actions$.pipe(
+
+  loadMyTeams$ = createEffect(() => this.actions$.pipe(
     ofType<LoadMyTeams>(TeamsActionTypes.LoadMyTeams),
     exhaustMap(() => this.authFacade.userId$),
     distinctUntilChanged(),
     switchMap((uid) => this.teamService.getUserTeams(uid).pipe(
       map(teams => new MyTeamsLoaded(teams.filter(t => t.members[0] !== '0'), uid))
     ))
-  );
+  ));
 
-  @Effect()
-  loadTeam$ = this.actions$.pipe(
+
+  loadTeam$ = createEffect(() => this.actions$.pipe(
     ofType<LoadTeam>(TeamsActionTypes.LoadTeam),
     withLatestFrom(this.teamsFacade.allTeams$),
     filter(([action, allTeams]) => {
@@ -34,33 +34,30 @@ export class TeamsEffects {
       );
     }),
     map(team => new TeamLoaded(<Team>team))
-  );
+  ));
 
-  @Effect()
-  createTeamInDatabase$ = this.actions$.pipe(
+
+  createTeamInDatabase$ = createEffect(() => this.actions$.pipe(
     ofType<CreateTeam>(TeamsActionTypes.CreateTeam),
     withLatestFrom(this.authFacade.userId$),
     switchMap(([action, userId]) => {
       action.payload.leader = userId;
       action.payload.members.push(userId);
       return this.teamService.add(action.payload);
-    }),
-    switchMap(() => EMPTY)
-  );
+    })
+  ), { dispatch: false });
 
-  @Effect()
-  updateTeamInDatabase$ = this.actions$.pipe(
+
+  updateTeamInDatabase$ = createEffect(() => this.actions$.pipe(
     ofType<UpdateTeam>(TeamsActionTypes.UpdateTeam),
-    switchMap(action => this.teamService.set(action.payload.$key, action.payload)),
-    switchMap(() => EMPTY)
-  );
+    switchMap(action => this.teamService.set(action.payload.$key, action.payload))
+  ), { dispatch: false });
 
-  @Effect()
-  deleteTeamFromDatabase$ = this.actions$.pipe(
+
+  deleteTeamFromDatabase$ = createEffect(() => this.actions$.pipe(
     ofType<DeleteTeam>(TeamsActionTypes.DeleteTeam),
-    switchMap(action => this.teamService.remove(action.payload)),
-    switchMap(() => EMPTY)
-  );
+    switchMap(action => this.teamService.remove(action.payload))
+  ), { dispatch: false });
 
   constructor(
     private actions$: Actions,
